moves annotation view setup into new file / class ItemAnnotationView
